+++ end_to_end_test/lib/serializers_nnbd.dart
@@ -36,6 +36,7 @@ part 'serializers_nnbd.g.dart';
   ConcreteGeneric,
   EnumWithInt,
   FallbackEnum,
+  FallbackNumberEnum,
   FieldDiscoveryValue,
   Fish,
   GenericContainer,